fix: standardize contentMapper content.config for collection list pages
