+++ src/pages/news/index.astro
@@ -5,7 +5,7 @@ import type { CollectionEntry } from "astro:content";
 import payloadFetch from "@/utilities/payload-fetch";
 import PagesSection from "@/components/PagesSection.astro";
 import CollectionItemList from "@/components/CollectionItemList.astro";
-import { newsMapper } from "@/utilities/newsMapper";
+import { newsMapper } from "@/utilities/contentMapper";
 
 
 const collectionName = "news";
