+++ src/pages/posts/index.astro
@@ -5,7 +5,7 @@ import CollectionItemList from "@/components/CollectionItemList.astro";
 import PagesSection from "@/components/PagesSection.astro";
 import type { CollectionEntry } from "astro:content";
 import payloadFetch from "@/utilities/payload-fetch";
-import { postsMapper } from "@/utilities/postsMapper";
+import { postsMapper } from "@/utilities/contentMapper";
 
 const collectionName = "posts";
 let posts: CollectionEntry<typeof collectionName>["data"][];
