+++ src/pages/reports/[slug].astro
@@ -6,7 +6,7 @@ import RichText from "@/components/RichText.astro";
 import payloadFetch from "@/utilities/payload-fetch";
 import PagesSection from "@/components/PagesSection.astro";
 import Tags, { type Tag } from "@/components/Tags.astro";
-import { reportMapper } from "@/utilities/collections";
+import { reportMapper } from "@/utilities/contentMapper"
 import Media from "@/components/Media.astro";
 
 const { slug } = Astro.params;
